pyproject.toml
[tool.poetry]
name = "the-factory-bot"
version = "0.1.0"
description = ""
authors = ["vladimirmac <vladimirbodrin89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.7"
pytelegrambotapi = "^4.12.0"
django-redis = "^5.3.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
pytest-mock = "^3.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
multi_line_output = 1
include_trailing_comma = true
line_length = 120
combine_as_imports = true
combine_star = true
skip_glob = ['**/migrations/**','.git','.venv']

[tool.autoflake]
in-place = true
